backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
groq==0.26.0
h11==0.16.0
idna==3.11
openai==2.32.0
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
